+++ attributes/default.rb
@@ -1,5 +1,9 @@
-default['drone']['package_url'] = "http://downloads.drone.io/latest/drone.deb"
-default['drone']['temp_file'] = "/tmp/drone.deb"
+default['drone']['install_type'] = 'source'
+
+default['drone']['git_url'] = 'https://github.com/drone/drone.git'
+default['drone']['git_ref'] = 'v0.2.0'
+
+default['drone']['package_url'] = 'https://downloads.drone.io/latest/drone.deb'
 default['drone']['config_file'] = '/etc/init/drone.conf'
 default['drone']['droned_opts'] = '--port=:80'
 default['drone']['drone_tmp'] = '/tmp/drone'