+++ metadata.rb
@@ -7,5 +7,8 @@ long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.4.0'
 
 depends 'docker'
+depends 'golang'
 
+supports 'redhat', '>= 6.5'
+supports 'centos', '>= 6.5'
 supports 'ubuntu', '>= 12.04'